requirements.txt
torch==2.2.0 --index-url https://download.pytorch.org/whl/nightly/cu118
torchvision==0.17.0 --index-url https://download.pytorch.org/whl/nightly/cu118
tensorboardX==2.6.2.2
tensorboard==2.8.0
opencv-python==4.8.1.78
scipy==1.12.0
numpy==1.24.1
PyYAML==5.4
tqdm==4.66.1
